fix: rename best of to teams in backend
