+++ clients/web/src/lib/components/molecules/scrims/fixtures/index.ts
@@ -1,3 +1,3 @@
 export {default as RoundRobinFixture} from "./RoundRobinFixture.svelte";
-export {default as BestOfFixture} from "./BestOfFixture.svelte";
+export {default as TeamsFixture} from "./TeamsFixture.svelte";
 export {default as GameCard} from "./GameCard.svelte";
